requirements.txt
fastapi==0.115.9
uvicorn==0.34.0
transformers==4.51.3
sentence-transformers==4.1.0
langchain==0.3.24
faiss-cpu==1.7.2  
openai==1.76.0
chromadb==1.0.7
python-dotenv==1.1.0
nltk==3.9.1
numpy==1.26.4
pdfplumber==0.11.6
google-generativeai==0.8.3  
underthesea==6.8.4  